CHG: Move Selector to distinct module
- base_data is long enough already
- Selector actually also depends on definitions done in derived classes

Changes to be committed:
	modified:   syncopy/datatype/base_data.py
	modified:   syncopy/datatype/continuous_data.py
	modified:   syncopy/datatype/discrete_data.py
	new file:   syncopy/datatype/selector.py
	modified:   syncopy/tests/test_computationalroutine.py
	modified:   syncopy/tests/test_continuousdata.py
	modified:   syncopy/tests/test_selectdata.py
	modified:   syncopy/tests/test_specest.py
